rename files according to new naming convention
